+++ test_syscall_exec.go
@@ -1,0 +1,15 @@
+// +build windows
+
+package main
+
+import (
+	"fmt"
+	"syscall"
+)
+
+func main() {
+	err := syscall.Exec("cmd.exe", []string{"cmd.exe"}, []string{})
+	if err != nil {
+		fmt.Printf("syscall.Exec error: %v\n", err)
+	}
+}